pyproject.toml
[tool.poetry]
name = "drfproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
skip-string-normalization = true
exclude = '''
/(
  migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 119